[repo groups] "Don't show again" for deletion confirmation
Introduce a preference for showing this confirmation dialog (by
default true). Add a checkbox to the dialog to toggle the setting.
Add an item on the "Confirmations & Warnings" preference page to
set and re-set the setting.

Show the dialog only if the setting is true, and there are any
repositories in the to-be-deleted repository groups.

+++ org.eclipse.egit.ui/src/org/eclipse/egit/ui/internal/repository/tree/command/DeleteRepositoryGroupCommand.java
@@ -15,14 +15,17 @@ import java.util.List;
 
 import org.eclipse.core.commands.ExecutionEvent;
 import org.eclipse.core.commands.ExecutionException;
+import org.eclipse.egit.ui.Activator;
+import org.eclipse.egit.ui.UIPreferences;
 import org.eclipse.egit.ui.internal.groups.RepositoryGroup;
 import org.eclipse.egit.ui.internal.groups.RepositoryGroups;
 import org.eclipse.egit.ui.internal.repository.tree.RepositoryGroupNode;
 import org.eclipse.egit.ui.internal.repository.tree.RepositoryTreeNode;
+import org.eclipse.jface.preference.IPreferenceStore;
 import org.eclipse.jface.window.Window;
 
 /**
- * Deletes a repository group, the repositories themselves are not affeced
+ * Deletes a repository group, the repositories themselves are not affected.
  */
 public class DeleteRepositoryGroupCommand
 		extends RepositoriesViewCommandHandler<RepositoryTreeNode> {
@@ -32,17 +35,32 @@ public class DeleteRepositoryGroupCommand
 		List<RepositoryGroup> groupsToDelete = new ArrayList<>();
 		List<RepositoryGroupNode> groupsNodes = new ArrayList<>();
 		List<RepositoryTreeNode> elements = getSelectedNodes();
+		int numberOfRepos = 0;
 		for (Object element : elements) {
 			if (element instanceof RepositoryGroupNode) {
 				RepositoryGroupNode groupNode = (RepositoryGroupNode) element;
 				groupsNodes.add(groupNode);
-				groupsToDelete.add(groupNode.getObject());
+				RepositoryGroup group = groupNode.getObject();
+				groupsToDelete.add(group);
+				numberOfRepos += group.getRepositoryDirectories().size();
 			}
 		}
 		if (!groupsToDelete.isEmpty()) {
-			DeleteRepositoryGroupConfirmDialog confirmDelete = new DeleteRepositoryGroupConfirmDialog(
-					getShell(event), groupsNodes);
-			if (confirmDelete.open() == Window.OK) {
+			IPreferenceStore store = Activator.getDefault()
+					.getPreferenceStore();
+
+			boolean delete = true;
+			if (numberOfRepos > 0 && store
+					.getBoolean(UIPreferences.SHOW_DELETE_REPO_GROUP_WARNING)) {
+				DeleteRepositoryGroupConfirmDialog confirmDelete = new DeleteRepositoryGroupConfirmDialog(
+						getShell(event), groupsNodes);
+				delete = confirmDelete.open() == Window.OK;
+				if (!confirmDelete.showAgain()) {
+					store.setValue(UIPreferences.SHOW_DELETE_REPO_GROUP_WARNING,
+							false);
+				}
+			}
+			if (delete) {
 				RepositoryGroups.getInstance().delete(groupsToDelete);
 				getView(event).refresh();
 			}
